Add CC compatible coroutine API
CC functions which interact with the MC world cause any coroutine to abrutly stop and be unrecoverable. The generator API provides a friendly consumer and performant API to execute a function outside of the coroutine. A full coroutine-like API can be used or for more common scenarios a simply 'run' function which can be used like an iterator